@@ -1,36 +1,41 @@
 {
-    "name": "frontend",
-    "dependencies": {
-        "react": "^15.5.4",
-        "react-dom": "^15.5.4"
-    },
-    "version": "0.1.0",
-    "devDependencies": {
-        "autoprefixer": "^7.1.1",
-        "babel-core": "^6.24.1",
-        "babel-eslint": "^7.2.3",
-        "babel-loader": "^7.0.0",
-        "babel-preset-react-app": "^2.2.0",
-        "css-loader": "^0.28.1",
-        "eslint": "^3.19.0",
-        "eslint-config-airbnb": "^15.0.1",
-        "eslint-loader": "^1.7.1",
-        "eslint-plugin-import": "^2.2.0",
-        "eslint-plugin-jsx-a11y": "^5.0.1",
-        "eslint-plugin-react": "^7.0.1",
-        "extract-text-webpack-plugin": "^2.1.0",
-        "file-loader": "^0.11.1",
-        "html-webpack-plugin": "^2.28.0",
-        "json-loader": "^0.5.4",
-        "postcss-flexbugs-fixes": "^3.0.0",
-        "postcss-loader": "^2.0.5",
-        "react-dev-utils": "^1.0.0",
-        "style-loader": "^0.17.0",
-        "webpack": "^2.5.1",
-        "webpack-dev-server": "^2.4.5"
-    },
-    "scripts": {
-        "start": "node scripts/start.js",
-        "build": "export NODE_ENV=production && rm -rf ./build && webpack --env.production --optimize-minimize"
-    }
+  "name": "fake",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/user-event": "^14.5.2",
+    "axios": "^1.7.9",
+    "cra-template": "1.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^7.1.1",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^4.2.4"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
